[INT] Moved directives into a separate folder
