+++ shell/directives/focus.js
