+++ shell/initialize/index.js
@@ -8,7 +8,7 @@ import App from './App.js';
 import { setContext, getLocation, getRouteData, normalizeError } from '../utils/nuxt';
 import { createStore } from '../config/store.js';
 import { UPGRADED, _FLAGGED, _UNFLAG } from '@shell/config/query-params';
-import { loadDirectives } from '@shell/plugins';
+import { loadDirectives } from '@shell/directives/index';
 import { installPlugins } from '@shell/initialize/plugins';
 
 // Prevent extensions from overriding existing directives
